requirements.txt
amqp==5.1.1
async-timeout==4.0.2
pyahocorasick==2.1.0
base58==2.1.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.0
Deprecated==1.2.13
ecdsa==0.15
eth-abi==2.2.0
eth-hash==0.3.3
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.2.2
gunicorn==20.1.0
h11==0.13.0
httpcore==0.12.3
httpx==0.16.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4
pymysql==1.0.2
MarkupSafe==2.1.1
packaging==21.3
parsimonious==0.8.1
prometheus-client==0.16.0
prompt-toolkit==3.0.30
pycryptodome==3.15.0
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
toolz==0.12.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
web3==5.31.3
cryptography==41.0.5
